+++ core/src/main/scala/mongo4s/MongoDatabase.scala
@@ -19,10 +19,13 @@ trait MongoDatabase[F[*], S[*]]:
       naming: FieldNaming = FieldNaming.identity,
   )(using BsonDocumentCodec[A]): F[MongoCollection[F, S, A]]
 
-  def getDirectCollection[A](
+  def getDirectCollection[A](using
+      codec: WireCodec[A],
+      tag: ClassTag[A]
+  )(
       collectionName: String,
-      naming: FieldNaming = FieldNaming.identity,
-  )(using WireCodec[A], ClassTag[A]): F[MongoCollection[F, S, A]]
+      naming: FieldNaming = codec.fieldNaming,
+  ): F[MongoCollection[F, S, A]]
 
   def listCollectionNames(using session: Option[ClientSession] = None)(using Streamable[S, String]): S[String]
   def listCollections(using session: Option[ClientSession] = None)(using Streamable[S, BsonDocument]): S[BsonDocument]